# remove double bound check from StringBuilder.Append(char)